globals gravity, player flap and clamp to sides

+++ Transform.pde
@@ -1,9 +1,9 @@
 class Transform {
-  int x;
-  int y;
-  int r;
+  float x;
+  float y;
+  float r;
 
-  Transform(int tempX, int tempY, int tempR){
+  Transform(float tempX, float tempY, float tempR){
     x = tempX;
     y = tempY;
     r = tempR;
